# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: InsuraBridge - Insurance Middleware Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: userId
          in: query
          required: false
          schema:
            type: string
        - name: unreadOnly
          in: query
          required: false
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create a notification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationRequest"
      responses:
        "201":
          description: Notification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/read-all:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read for a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: integer
              required:
                - userId
      responses:
        "200":
          description: All notifications marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # CLAIMS
...
  /scrutiny:
    get:
      operationId: listScrutiny
      tags: [scrutiny]
      summary: List scrutiny/settlement cases
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: claimId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of scrutiny cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScrutinyCase"
    post:
      operationId: createScrutiny
      tags: [scrutiny]
      summary: Create a scrutiny case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateScrutinyRequest"
      responses:
        "201":
          description: Scrutiny case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrutinyCase"
  /scrutiny/{id}:
    patch:
      operationId: updateScrutiny
      tags: [scrutiny]
      summary: Update scrutiny case
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScrutinyRequest"
      responses:
        "200":
          description: Scrutiny case updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScrutinyCase"

  # PORTABILITY
  /portability:
    get:
      operationId: listPortability
      tags: [portability]
      summary: List portability requests
      responses:
        "200":
          description: List of portability requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PortabilityRequest"
    post:
      operationId: createPortability
      tags: [portability]
      summary: Submit a portability request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePortabilityRequest"
      responses:
        "201":
          description: Portability request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortabilityRequest"
  /portability/{id}:
    patch:
      operationId: updatePortability
      tags: [portability]
      summary: Update portability request status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePortabilityRequest"
      responses:
        "200":
          description: Portability request updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PortabilityRequest"

  # RENEWALS
  /renewals:
    get:
      operationId: listRenewals
      tags: [renewals]
      summary: List policy renewals
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of renewals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Renewal"
    post:
      operationId: createRenewal
      tags: [renewals]
      summary: Create a renewal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRenewalRequest"
      responses:
        "201":
          description: Renewal created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Renewal"
  /renewals/{id}:
    patch:
      operationId: updateRenewal
      tags: [renewals]
      summary: Update renewal status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRenewalRequest"
      responses:
        "200":
          description: Renewal updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Renewal"

  # MEMBERS
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [message, claim_update, document, bill, settlement, system]
        relatedId:
          type: integer
          nullable: true
        relatedType:
          type: string
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, title, message, type, isRead, createdAt]
    CreateNotificationRequest:
      type: object
      properties:
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [message, claim_update, document, bill, settlement, system]
        relatedId:
          type: integer
          nullable: true
        relatedType:
          type: string
          nullable: true
      required: [userId, title, message, type]
...
    ScrutinyCase:
      type: object
      properties:
        id:
          type: integer
        claimId:
          type: integer
          nullable: true
        claimNumber:
          type: string
          nullable: true
        patientName:
          type: string
        billAmount:
          type: number
        scrutinizedAmount:
          type: number
          nullable: true
        deductions:
          type: number
          nullable: true
        deductionReasons:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [pending, in_review, completed, escalated]
        assignedTo:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, patientName, billAmount, status, deductionReasons, createdAt]
    CreateScrutinyRequest:
      type: object
      properties:
        claimId:
          type: integer
          nullable: true
        claimNumber:
          type: string
          nullable: true
        patientName:
          type: string
        billAmount:
          type: number
        assignedTo:
          type: string
          nullable: true
        remarks:
          type: string
          nullable: true
      required: [patientName, billAmount]
    UpdateScrutinyRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_review, completed, escalated]
        scrutinizedAmount:
          type: number
          nullable: true
        deductions:
          type: number
          nullable: true
        deductionReasons:
          type: array
          items:
            type: string
        remarks:
          type: string
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
    PortabilityRequest:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        fromInsurerName:
          type: string
        toInsurerName:
          type: string
        policyNumber:
          type: string
        sumInsured:
          type: number
        newSumInsured:
          type: number
          nullable: true
        portabilityReason:
          type: string
          nullable: true
        previousClaimHistory:
          type: string
          nullable: true
        status:
          type: string
          enum: [initiated, under_review, approved, rejected, completed]
        requestedAt:
          type: string
          format: date-time
        effectiveDate:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, customerId, customerName, fromInsurerName, toInsurerName, policyNumber, sumInsured, status, requestedAt, createdAt]
    CreatePortabilityRequest:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        fromInsurerName:
          type: string
        toInsurerName:
          type: string
        policyNumber:
          type: string
        sumInsured:
          type: number
        newSumInsured:
          type: number
          nullable: true
        portabilityReason:
          type: string
          nullable: true
        previousClaimHistory:
          type: string
          nullable: true
        requestedAt:
          type: string
          format: date-time
      required: [customerId, customerName, fromInsurerName, toInsurerName, policyNumber, sumInsured, requestedAt]
    UpdatePortabilityRequest:
      type: object
      properties:
        status:
          type: string
          enum: [initiated, under_review, approved, rejected, completed]
        effectiveDate:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
    Renewal:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        policyNumber:
          type: string
        insurerName:
          type: string
        expiryDate:
          type: string
          format: date-time
        renewalDate:
          type: string
          format: date-time
          nullable: true
        currentSumInsured:
          type: number
        newSumInsured:
          type: number
          nullable: true
        currentPremium:
          type: number
        newPremium:
          type: number
          nullable: true
        status:
          type: string
          enum: [pending, initiated, payment_due, completed, lapsed]
        memberCount:
          type: integer
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, customerId, customerName, policyNumber, insurerName, expiryDate, currentSumInsured, currentPremium, status, memberCount, createdAt]
    CreateRenewalRequest:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        policyNumber:
          type: string
        insurerName:
          type: string
        expiryDate:
          type: string
          format: date-time
        currentSumInsured:
          type: number
        newSumInsured:
          type: number
          nullable: true
        currentPremium:
          type: number
        newPremium:
          type: number
          nullable: true
        memberCount:
          type: integer
        notes:
          type: string
          nullable: true
      required: [customerId, customerName, policyNumber, insurerName, expiryDate, currentSumInsured, currentPremium, memberCount]
    UpdateRenewalRequest:
      type: object
      properties:
        status:
          type: string
          enum: [pending, initiated, payment_due, completed, lapsed]
        renewalDate:
          type: string
          format: date-time
          nullable: true
        newSumInsured:
          type: number
          nullable: true
        newPremium:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
...