# lib/api-zod/src/generated/types/createScrutinyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * InsuraBridge - Insurance Middleware Platform API
 * OpenAPI spec version: 0.2.0
 */

export interface CreateScrutinyRequest {
  claimId?: number | null;
  claimNumber?: string | null;
  patientName: string;
  billAmount: number;
  assignedTo?: string | null;
  remarks?: string | null;
}